successful death of docker pod

+++ src/dctl-contract-test.js
@@ -28,17 +28,19 @@ program
       if (!directoryExists(localSecrets)) throw new Error(`Missing secrets directory "${localSecrets}". Try running 'dctl contract init --help'.`);
       if (!fileExists(localEnv)) throw new Error(`Missing .env file ${localEnv}. Try running 'dctl contract init --help'.`);
       const config = JSON.parse(await fs.promises.readFile(path.join(testRoot, 'config.json'), 'utf-8'));
+      console.log(`${config.webserverImage} ${config.webServerStartCommand}`)
       const webserverStartCommand = `docker run \
-      -l env=dragonchain_test_env
+      -d \
+      -l env=dragonchain_test_env \
       -p 38404:8080 \
       --env-file ${localEnv} \
       --entrypoint '' \
-      ${config.webserverImage} ${webserverStartCommand}
+      ${config.webserverImage} ${config.webServerStartCommand}
       `;
       shell.exec(webserverStartCommand);
       console.log('Booting fake webserver pod');
       const command = `printf '${transaction(payload)}' | docker run -i \
-      --rm
+      --rm \
       -v ${localHeap}:${remoteHeap} \
       -v ${localSecrets}:${remoteSecrets}:ro \
       --env-file ${localEnv} \
@@ -121,9 +123,10 @@ async function directoryExists(dirPath) {
  */
 function killWebserver() {
     try {
-        const psCommand = 'docker ps --filter env=dragonchain_test_env --format "{{.ID}}"'; // returns only the ID from containers that match this label
+        const psCommand = 'docker ps --filter label=env=dragonchain_test_env --format "{{.ID}}"'; // returns only the ID from containers that match this label
         const imageIds = shell.exec(psCommand).trim().split('\n'); // array of docker image IDs
-        imageIds.forEach(id => (shell.exec(`docker stop ${id} && docker rm ${id}`))); // remove said docker images
+        console.log(`FOUND imageIds: ${imageIds}`);
+        if (imageIds.length > 0) imageIds.forEach(id => (shell.exec(`docker stop ${id} && docker rm ${id}`))); // remove said docker images
         return true;
     } catch (error) {
         console.error(`Error occurred while trying to remove docker images: ${error}`);
